moved types + fixed get localStorage
