+++ src/app/features/contentfull/mocks/diploma.mock.ts
