+++ src/app/features/contentfull/mocks/index.ts
@@ -1,1 +1,8 @@
+export * from './asset.mock';
+export * from './contentfull.service.mock';
+export * from './diploma.mock';
+export * from './job.mock';
+export * from './place.mock';
 export * from './rich-text.mock';
+export * from './skill.mock';
+export * from './training.mock';
