+++ src/app/features/contentfull/types/diploma/diploma.type.spec.ts
