+++ src/app/features/contentfull/types/index.ts
@@ -1,4 +1,8 @@
+export * from './asset/asset.type';
+export * from './diploma/diploma.type';
 export { entrySchema } from './entry.type';
+export * from './job/job.type';
+export * from './place/place.type';
 export {
 	formattedRichTextSchema,
 	isFormattedRichText,
@@ -6,3 +10,5 @@ export {
 	type FormattedRichText,
 	type RichTextNode,
 } from './rich-text/rich-text.type';
+export * from './skill/skill.type';
+export * from './training/training.type';
