+++ src/app/features/cv/components/skill/skills-list.component.ts
@@ -5,7 +5,7 @@ import {
 	computed,
 	input,
 } from '@angular/core';
-import type { Skill } from '../../types';
+import type { Skill } from '@feat/contentfull/types';
 import { SkillPillComponent } from './skill-pill.component';
 
 @Component({
