+++ src/app/features/cv/mocks/cv-event.mock.ts
@@ -1,6 +1,6 @@
+import { validJobInput } from '../../contentfull/mocks/job.mock';
+import { validTrainingInput } from '../../contentfull/mocks/training.mock';
 import type { CvEventInput } from '../types/cv-event/cv-event.type';
-import { validJobInput } from './job.mock';
-import { validTrainingInput } from './training.mock';
 
 export const validCvEventInput = (type: 'job' | 'training'): CvEventInput =>
-	type === 'job' ? validJobInput : validTrainingInput;
+	type === 'job' ? validJobInput() : validTrainingInput();
