+++ src/app/features/cv/mocks/index.ts
@@ -1,6 +1,1 @@
-export * from './asset.mock';
-export * from './diploma.mock';
-export * from './job.mock';
-export * from './place.mock';
-export * from './skill.mock';
-export * from './training.mock';
+export * from './cv-event.mock';
