+++ src/app/features/cv/types/index.ts
@@ -1,18 +1,16 @@
-export { diplomaSchema, type Diploma } from './diploma/diploma.type';
-export {
-	isJob,
-	isJobInput,
-	Job,
-	jobSchema,
-	type RemotePolicy,
-} from './job/job.type';
-export { placeSchema, type Place } from './place/place.type';
+export { isJob, Job } from './job/job.type';
 
 export {
+	contractTypesKeys,
 	EVENT_TYPES,
 	type ContractType,
 	type CvEventType,
 } from './cv-event/cv-event-type.type';
-export { CvEvent, type CvEventField } from './cv-event/cv-event.type';
-export { isSkill, skillSchema, type Skill } from './skill/skill.type';
-export { isTraining, Training, trainingSchema } from './training/training.type';
+
+export {
+	CvEvent,
+	cvEventSchema,
+	type CvEventField,
+} from './cv-event/cv-event.type';
+
+export { isTraining, Training } from './training/training.type';
