+++ src/app/features/cv/types/job/job.type.ts
@@ -1,44 +1,17 @@
-import type { FormattedRichText } from '@feat/contentfull/types';
-import { entrySchema, formattedRichTextSchema } from '@feat/contentfull/types';
+import type { FormattedRichText, Place, Skill } from '@feat/contentfull/types';
 
 import { isSchemaType } from '@shared/functions';
 import { z } from 'zod';
-import { assetSchema } from '../asset/asset.type';
-import type { ContractType } from '../cv-event/cv-event-type.type';
-import { contractTypesKeys } from '../cv-event/cv-event-type.type';
-import { CvEvent, cvEventSchema } from '../cv-event/cv-event.type';
-import type { Place } from '../place/place.type';
-import { placeSchema } from '../place/place.type';
-import type { Skill } from '../skill/skill.type';
-import { skillSchema } from '../skill/skill.type';
-
-export const REMOTE_POLICIES = [
-	{ value: 'à distance', label: 'À distance' },
-	{ value: 'hybride', label: 'Hybride' },
-	{ value: 'sur site', label: 'sur site' },
-] as const;
-export type RemotePolicy = (typeof REMOTE_POLICIES)[number]['value'];
-const RemotePolicyKeys = REMOTE_POLICIES.map(p => p.value) as [
-	RemotePolicy,
-	...RemotePolicy[],
-];
-
-// JOB
-export const jobSchema = entrySchema.extend({
-	company: placeSchema,
-	remotePolicy: z.enum(RemotePolicyKeys),
-	contractType: z.enum(contractTypesKeys),
-	title: z.string(),
-	summary: z.string(),
-	description: formattedRichTextSchema, // tasks
-	startDate: z.coerce.date(),
-	endDate: z.coerce.date().nullish(),
-	assets: z.array(assetSchema).nullish(),
-	skills: z.array(skillSchema),
-});
-export type JobEntry = z.infer<typeof jobSchema>;
-export const isJobInput = (entry: unknown): entry is JobEntry =>
-	isSchemaType(entry, jobSchema, 'JobInput');
+import {
+	RemotePolicyKeys,
+	type JobEntry,
+	type RemotePolicy,
+} from '../../../contentfull/types/job/job.type';
+import type { ContractType } from '../../../cv/types/cv-event/cv-event-type.type';
+import {
+	CvEvent,
+	cvEventSchema,
+} from '../../../cv/types/cv-event/cv-event.type';
 
 export class Job extends CvEvent {
 	readonly type: ContractType;
